+++ src/types/json.rs
@@ -176,7 +176,7 @@ where
     fn from_request(req: &HttpRequest, payload: &mut Payload) -> Self::Future {
         let req2 = req.clone();
         let (limit, err) = req
-            .route_data::<JsonConfig>()
+            .app_data::<JsonConfig>()
             .map(|c| (c.limit, c.ehandler.clone()))
             .unwrap_or((32768, None));
 
@@ -220,17 +220,16 @@ where
 ///
 /// fn main() {
 ///     let app = App::new().service(
-///         web::resource("/index.html").route(
-///             web::post().data(
-///                 // change json extractor configuration
-///                 web::Json::<Info>::configure(|cfg| {
-///                     cfg.limit(4096)
-///                        .error_handler(|err, req| {  // <- create custom error response
-///                             error::InternalError::from_response(
-///                                 err, HttpResponse::Conflict().finish()).into()
-///                        })
-///                 }))
-///                 .to(index))
+///         web::resource("/index.html").data(
+///             // change json extractor configuration
+///             web::Json::<Info>::configure(|cfg| {
+///                 cfg.limit(4096)
+///                    .error_handler(|err, req| {  // <- create custom error response
+///                         error::InternalError::from_response(
+///                             err, HttpResponse::Conflict().finish()).into()
+///                    })
+///             }))
+///             .route(web::post().to(index))
 ///     );
 /// }
 /// ```
@@ -431,7 +430,7 @@ mod tests {
                 header::HeaderValue::from_static("16"),
             )
             .set_payload(Bytes::from_static(b"{\"name\": \"test\"}"))
-            .route_data(JsonConfig::default().limit(10).error_handler(|err, _| {
+            .data(JsonConfig::default().limit(10).error_handler(|err, _| {
                 let msg = MyObject {
                     name: "invalid request".to_string(),
                 };
@@ -483,7 +482,7 @@ mod tests {
                 header::HeaderValue::from_static("16"),
             )
             .set_payload(Bytes::from_static(b"{\"name\": \"test\"}"))
-            .route_data(JsonConfig::default().limit(10))
+            .data(JsonConfig::default().limit(10))
             .to_http_parts();
 
         let s = block_on(Json::<MyObject>::from_request(&req, &mut pl));
@@ -500,7 +499,7 @@ mod tests {
                 header::HeaderValue::from_static("16"),
             )
             .set_payload(Bytes::from_static(b"{\"name\": \"test\"}"))
-            .route_data(
+            .data(
                 JsonConfig::default()
                     .limit(10)
                     .error_handler(|_, _| JsonPayloadError::ContentType.into()),