+++ src/web.rs
@@ -15,7 +15,7 @@ use crate::scope::Scope;
 use crate::service::WebService;
 
 pub use crate::config::ServiceConfig;
-pub use crate::data::{Data, RouteData};
+pub use crate::data::Data;
 pub use crate::request::HttpRequest;
 pub use crate::types::*;